Move parse_case into its own file
